Iss: netcdf tool meta edit: file process draft

+++ hs_app_tools_netCDF/views.py
@@ -11,6 +11,8 @@ from django.core.files import File
 from django.core.files.base import ContentFile
 from django.core.urlresolvers import reverse
 import numpy as np
+from hs_core import hydroshare
+from hs_core.hydroshare import utils
 
 # views for index page
 def index_view(request, shortkey, state):
@@ -54,7 +56,7 @@ def meta_edit_view(request, shortkey, **kwargs):
         meta_elements = request.POST.getlist('meta_elements')
         file_process = request.POST.getlist('file_process')
 
-        if meta_elements and file_process:
+        if meta_elements:
             check_info = run_meta_edit_tool(res, meta_elements, file_process)
             if check_info:
                 messages.add_message(request, messages.ERROR, check_info)
@@ -62,13 +64,14 @@ def meta_edit_view(request, shortkey, **kwargs):
                 messages.add_message(request, messages.SUCCESS, 'Success! Metadata Editing is finished.')
         else:
             messages.add_message(request, messages.ERROR,
-                                 'Error! Metadata Editing: please select both of the required checkboxes.')
+                                 'Error! Metadata Editing: please select the metadata elements.')
 
     return HttpResponseRedirect(reverse('nc_tools:index', args=[res.short_id, 'processing']))
 
 
 def run_meta_edit_tool(res, meta_elements, file_process):
     check_info = ''
+    import os
     for f in ResourceFile.objects.filter(object_id=res.id):
         ext = os.path.splitext(f.resource_file.name)[-1]
         if ext == '.nc':
@@ -77,11 +80,17 @@ def run_meta_edit_tool(res, meta_elements, file_process):
             if not nc_tools_obj:
                 nc_tools_obj = NetcdfTools(short_id=res.short_id)
             else:
-                nc_tools_obj.meta_edit_file.delete()
+                if nc_tools_obj.meta_edit_file:
+                    storage, path = nc_tools_obj.meta_edit_file.storage, nc_tools_obj.meta_edit_file.path
+                    nc_tools_obj.meta_edit_file.delete()
+                    storage.delete(path)
+
 
 
             # add meta_edit initial file
             meta_edit_file_name = os.path.basename(f.resource_file.name)
+            if 'metaEdit_' not in os.path.basename(f.resource_file.name):
+                meta_edit_file_name = 'metaEdit_'+os.path.basename(f.resource_file.name)
             meta_edit_file = open(f.resource_file.file.name)  # ContentFile(f.resource_file.file.read())
             nc_tools_obj.meta_edit_file.save(meta_edit_file_name, File(meta_edit_file))
 
@@ -93,6 +102,51 @@ def run_meta_edit_tool(res, meta_elements, file_process):
             if check_info:
                 return check_info
             else:
+                if 'new_ver_res' in file_process:
+
+                    try:
+                        # add new .nc file
+
+                        for f in ResourceFile.objects.filter(object_id=res.id):
+                            f.resource_file.delete()
+                            f.delete()
+                        f = ResourceFile.objects.create(content_object=res)
+                        meta_edit_file_name = os.path.basename(nc_tools_obj.meta_edit_file.name).replace("metaEdit_",'')
+                        meta_edit_file = open(nc_tools_obj.meta_edit_file.file.name)
+                        f.resource_file.save(meta_edit_file_name, File(meta_edit_file))
+
+                        # create InMemoryUploadedFile text file to store the dump info and add it to resource
+
+                        import hs_app_netCDF.nc_functions.nc_dump as nc_dump
+                        if nc_dump.get_nc_dump_string_by_ncdump(nc_file_name):
+                            dump_str = nc_dump.get_nc_dump_string_by_ncdump(nc_file_name)
+                        else:
+                            dump_str = nc_dump.get_nc_dump_string(nc_file_name)
+                        if dump_str:
+                            from django.core.files.uploadedfile import InMemoryUploadedFile
+                            import StringIO, os
+                            io = StringIO.StringIO()
+                            io.write(dump_str)
+                            dump_file_name = '.'.join(os.path.basename(nc_file_name.replace("metaEdit_",'')).split('.')[:-1])+'_header_info.txt'
+                            dump_file = InMemoryUploadedFile(io, None, dump_file_name, 'text', io.len, None)
+                            dump_file.seek(0)
+                            hydroshare.add_resource_files(res.short_id, dump_file)
+
+                    except:
+                        for f in ResourceFile.objects.filter(object_id=res.id):
+                            if not f.resource_file:
+                                f.delete()
+                        check_info = "Error! Metadata Editing failed to create new version of resource with the edited file."
+
+                if 'new_res' in file_process:
+                    try:
+                        pass
+                    except:
+                        if check_info:
+                            check_info = "Error! Metadata Editing failed to create new version of resource and the new resource."
+                        else:
+                            check_info = "Error! Metadata Editing failed to create a new resource with the edited file."
+
                 return check_info
 
     check_info = "Error! Metadata Editing: there is no .nc file in the resource. "
@@ -113,20 +167,20 @@ def edit_meta_in_file(res, nc_file_name, meta_elements):
         if 'description' in meta_elements and res.metadata.description.abstract:  # Done
             nc_dataset.summary = res.metadata.description.abstract
 
-        if 'subjects' in meta_elements and res.metadata.subjects.all().first():  # Done
+        if 'subjects' in meta_elements and res.metadata.subjects.all().first():
             res_meta_subjects = []
             for res_subject in res.metadata.subjects.all():
                 res_meta_subjects.append(res_subject.value)
             nc_dataset.keywords = ','.join(res_meta_subjects)
 
-        if 'rights' in meta_elements and res.metadata.rights: #Done
-            nc_dataset.license = "{0} \n {1}".format(res.metadata.rights.statement, res.metadata.rights.url)
+        if 'rights' in meta_elements and res.metadata.rights:
+            nc_dataset.license = "{0} {1}".format(res.metadata.rights.statement, res.metadata.rights.url)
 
-        if 'publisher' in meta_elements: #Done TODO need to confirm about this element info
+        if 'publisher' in meta_elements: #TODO need to confirm about this element info
             nc_dataset.publisher_name = 'HydroShare'
             nc_dataset.publisher_url = 'http://www.hydroshare.org'
 
-        if 'identifier' in meta_elements and res.metadata.identifiers.all():  # Done
+        if 'identifier' in meta_elements and res.metadata.identifiers.all():
             if res.metadata.identifiers.all().filter(name="doi"):
                 res_identifier = res.metadata.identifiers.all().filter(name="doi")[0]
             else:
@@ -151,7 +205,6 @@ def edit_meta_in_file(res, nc_file_name, meta_elements):
             for variable in res.metadata.variables.all():
                 if variable.name in res_var_dict.keys():
                     res_var = res_var_dict[variable.name]
-                    b=res_var.__dict__.keys()
                     if variable.unit != 'Unknown':
                         res_var.setncattr('units', variable.unit)
                     if variable.descriptive_name:
@@ -178,7 +231,6 @@ def edit_meta_in_file(res, nc_file_name, meta_elements):
             nc_dataset.geospatial_lon_min = res_meta_cov.value['westlimit']
             nc_dataset.geospatial_lon_max = res_meta_cov.value['eastlimit']
 
-        res.metadata.contributors.all()
         if 'contributor' in meta_elements and res.metadata.contributors.all():
             res_contri_name = []
             for contributor in res.metadata.contributors.all():
@@ -193,12 +245,13 @@ def edit_meta_in_file(res, nc_file_name, meta_elements):
             if creator.description or creator.homepage:
                 nc_dataset.creator_url = creator.homepage if creator.homepage else creator.description
 
-
-
         # edit convention meta if ACDD terms are edited in file:
         if meta_elements:
             ori_con = nc_dataset.Conventions if hasattr(nc_dataset, "Conventions") else ''
-            nc_dataset.Conventions = ', '.join([ori_con, "ACDD-1.3"])
+            if not ori_con:
+                nc_dataset.Conventions = "ACDD-1.3"
+            elif "ACDD" not in ori_con:
+                nc_dataset.Conventions = ', '.join(ori_con.split(',')+["ACDD-1.3"])
 
         # close nc_dataset
         nc_dataset.close()
